Remove FetchSecret Task type
Description
===========

The task type `FetchSecrets` is a very opinioned piece of
code which should not belong in this plugin. I removed the class
and declared an abstract base type in it's place to set the
secret key and secrets resolver properties

Changes
=======

* ![REMOVE] `net.wooga.gradle.secrets.tasks.FetchSecrets`
* ![ADD] `net.wooga.gradle.secrets.tasks.SecretsTask`
